Modified Files: 	src/servers/mouse/main.cpp

+++ Mona/src/servers/mouse/main.cpp
@@ -158,8 +158,8 @@ void MouseServer::MessageLoop()
                 char x = (byte)(receive.arg2);
                 char y = (byte)(receive.arg3) * (-1);
 
-                this->posX += x * 1.5;
-                this->posY += y * 1.5;
+                this->posX += (int)(((double)x) * 1.5);
+                this->posY += (int)(((double)y) * 1.5);
 
                 /* mouse cursor size */
                 if (this->posX >= this->w) this->posX = this->w - 1;
